backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
moviepy==1.0.3
groq==0.3.0
langgraph==0.0.12
openai-whisper==20231117
werkzeug==3.0.1